delta4j/genea:     * Code flush.

+++ src/main/java/delta/genea/sql/ActSqlDriver.java
@@ -103,7 +103,7 @@ public class ActSqlDriver extends ObjectSqlDriver<Act>
       ResultSet rs=null;
       try
       {
-        _psGetByPrimaryKey.setLong(1,primaryKey);
+        _psGetByPrimaryKey.setLong(1,primaryKey.longValue());
         rs=_psGetByPrimaryKey.executeQuery();
         if (rs.next())
         {
